@@ -49,15 +49,14 @@
     "jest": "^20.0.3",
     "prettier": "^1.8.2",
     "raf": "^3.4.0",
-    "react": "^16.0.0",
-    "react-dom": "^16.0.0",
     "react-test-renderer": "^16.0.0",
     "webpack": "^1.13.2",
     "webpack-dev-server": "^1.16.1"
   },
-  "peerDependencies": {
-    "react": "^15.0.0-0 || ^16.0.0-0",
-    "react-dom": "^15.0.0-0 || ^16.0.0-0"
-  },
-  "sideEffects": false
+  "peerDependencies": {},
+  "sideEffects": false,
+  "dependencies": {
+    "react": "^16.5.1",
+    "react-dom": "^16.5.1"
+  }
 }
